Switch to nested shortcode
