+++ .eleventy.js
@@ -36,6 +36,8 @@ const Banner = require(`./${componentsDir}/Banner`);
 const Breadcrumbs = require(`./${componentsDir}/Breadcrumbs`);
 const CodelabsCallout = require(`./${componentsDir}/CodelabsCallout`);
 const Compare = require(`./${componentsDir}/Compare`);
+const Details = require(`./${componentsDir}/Details`);
+const DetailsSummary = require(`./${componentsDir}/DetailsSummary`);
 const Hero = require(`./${componentsDir}/Hero`);
 const Instruction = require(`./${componentsDir}/Instruction`);
 const PathCard = require(`./${componentsDir}/PathCard`);
@@ -43,7 +45,6 @@ const PostCard = require(`./${componentsDir}/PostCard`);
 const YouTube = require(`./${componentsDir}/YouTube`);
 
 const tagsDir = 'src/site/_includes/components/tags';
-const {Details} = require(`./${tagsDir}/Details`);
 const {Image, Figure} = require(`./${tagsDir}/Image`);
 
 const collectionsDir = 'src/site/_collections';
@@ -142,6 +143,8 @@ module.exports = function(config) {
   config.addShortcode('Breadcrumbs', Breadcrumbs);
   config.addShortcode('CodelabsCallout', CodelabsCallout);
   config.addPairedShortcode('Compare', Compare);
+  config.addPairedShortcode('Details', Details);
+  config.addPairedShortcode('DetailsSummary', DetailsSummary);
   config.addShortcode('Hero', Hero);
   config.addShortcode('Instruction', Instruction);
   config.addShortcode('PathCard', PathCard);
@@ -154,7 +157,6 @@ module.exports = function(config) {
   // CUSTOM TAGS
   //----------------------------------------------------------------------------
   config.addNunjucksTag('Image', Image);
-  config.addNunjucksTag('Details', Details);
   config.addNunjucksTag('Figure', Figure);
 
   // https://www.11ty.io/docs/config/#data-deep-merge
